mv get_duration to get_duration.py
