+++ radiko_timeshift_recorder/download.py
@@ -1,5 +1,4 @@
 import asyncio
-import json
 import logging
 import shlex
 import tempfile
@@ -32,30 +31,6 @@ def get_out_filepath(job: Job, out_dir: Path) -> Path:
     ).resolve()
 
     return trim_filestem(out_filepath)
-
-
-async def get_duration(filepath: Path) -> float:
-    proc = await asyncio.create_subprocess_exec(
-        "ffprobe",
-        "-hide_banner",
-        "-show_streams",
-        "-print_format",
-        "json",
-        str(filepath.resolve()),
-        stdout=asyncio.subprocess.PIPE,
-        stderr=asyncio.subprocess.PIPE,
-    )
-
-    stdout, stderr = await proc.communicate()
-
-    if proc.returncode != 0:
-        logger.debug(f"stdout: {stdout.decode().strip()}")
-        logger.debug(f"stderr: {stderr.decode().strip()}")
-        raise RuntimeError(
-            f"Failed to get duration of {filepath}: {stderr.decode().strip()}"
-        )
-
-    return float(json.loads(stdout)["streams"][0]["duration"])
 
 
 async def download_stream(url: str, out_filepath: Path) -> None:
